+++ bpm-engine-impl/src/main/java/io/takari/bpm/state/ProcessStatus.java
@@ -2,5 +2,5 @@ package io.takari.bpm.state;
 
 public enum ProcessStatus {
 
-    RUNNING, SUSPENDED, FINISHED, FAILED
+    RUNNING, SUSPENDED, FINISHED
 }